Build gems based on platform
The previous gem size peaked at over 250MB by providing an unnecessary
amount of unused binaries.

Using the os host(kernel name) and the architecture, separate out
individual platform gemspecs provide a subset of the needed files.

This change is NOT perfect. It can only limit platforms by kernel and
architecture. It does not take the distribution of Linux into account,
so it will still provide an unnecessary number of files, but fewer.

Minor updates:

- Set minimum ruby version
- Set versions for development dependencies